+++ 2017/src/main/java/de/adventofcode/chrisgw/day21/FractalArt.java
@@ -207,7 +207,7 @@ public class FractalArt {
     }
 
 
-    public int countLitPixels() {
+    public long countLitPixels() {
         return pixelGrid.countLitPixels();
     }
